requirements.txt
# LangChain Core and Extensions
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.9
langchain-text-splitters>=0.3.0
langchain-core>=0.3.0
langgraph>=0.2.0

# Vector Database and Embeddings
pinecone-client>=3.0.0
langchain-pinecone>=0.1.0

# Cloud Storage
boto3>=1.34.0

# API Framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
python-multipart>=0.0.9

# Document Processing
pymupdf>=1.24.0

# HTTP Client
requests>=2.32.0

# Environment and Configuration
python-dotenv>=1.0.0

# Data Analysis and Evaluation
pandas>=2.2.0
datasets>=2.14.0

# LLM Evaluation with RAGAS
ragas>=0.1.20,<0.3.0

# OpenAI API
openai>=1.50.0

# Authentication & User Database
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
libsql-client>=0.3.0